+++ src/systems/filecoin_vm/sysactors/init_actor.go
@@ -5,54 +5,11 @@ import actor "github.com/filecoin-project/specs/systems/filecoin_vm/actor"
 import vmr "github.com/filecoin-project/specs/systems/filecoin_vm/runtime"
 import exitcode "github.com/filecoin-project/specs/systems/filecoin_vm/runtime/exitcode"
 import util "github.com/filecoin-project/specs/util"
-import ipld "github.com/filecoin-project/specs/libraries/ipld"
 
 const (
 	Method_InitActor_Exec = actor.MethodPlaceholder + iota
 	Method_InitActor_GetActorIDForAddress
 )
-
-////////////////////////////////////////////////////////////////////////////////
-// Boilerplate
-////////////////////////////////////////////////////////////////////////////////
-type InvocOutput = vmr.InvocOutput
-type Runtime = vmr.Runtime
-type Bytes = util.Bytes
-type Serialization = util.Serialization
-
-var CheckArgs = actor.CheckArgs
-var ArgPop = actor.ArgPop
-var ArgEnd = actor.ArgEnd
-
-func _loadState(rt Runtime) (vmr.ActorStateHandle, InitActorState) {
-	h := rt.AcquireState()
-	stateCID := ipld.CID(h.Take())
-	if ipld.CID_Equals(stateCID, ipld.EmptyCID()) {
-		rt.AbortAPI("Actor state not initialized")
-	}
-	stateBytes := rt.IpldGet(ipld.CID(stateCID))
-	if stateBytes.Which() != vmr.Runtime_IpldGet_FunRet_Case_Bytes {
-		rt.AbortAPI("IPLD lookup error")
-	}
-	state, err := Deserialize_InitActorState(Serialization(stateBytes.As_Bytes()))
-	if err != nil {
-		rt.AbortAPI("State deserialization error")
-	}
-	return h, state
-}
-func Release(rt Runtime, h vmr.ActorStateHandle, st InitActorState) {
-	checkCID := actor.ActorSubstateCID(rt.IpldPut(st.Impl()))
-	h.Release(checkCID)
-}
-func UpdateRelease(rt Runtime, h vmr.ActorStateHandle, st InitActorState) {
-	newCID := actor.ActorSubstateCID(rt.IpldPut(st.Impl()))
-	h.UpdateRelease(newCID)
-}
-func (st *InitActorState_I) CID() ipld.CID {
-	panic("TODO")
-}
-
-////////////////////////////////////////////////////////////////////////////////
 
 func (a *InitActorCode_I) Constructor(rt Runtime) InvocOutput {
 	h := rt.AcquireState()
